+++ packages/react-together/src/components/Chat/MessageBody.tsx
@@ -1,0 +1,28 @@
+import { MessageBodyProps } from './types'
+
+function timeToLocale(timestamp: number) {
+  const date = new Date(timestamp)
+  return date.toLocaleTimeString([], { hour: '2-digit', minute: '2-digit' })
+}
+
+export default function MessageBody({
+  isMe,
+  senderId,
+  message,
+  sentAt,
+  formatTime = timeToLocale
+}: MessageBodyProps) {
+  return (
+    <div className={`rt-message-divider ${isMe ? 'isMe' : ''}`}>
+      <div className={`rt-message-border ${isMe ? 'isMe' : ''}`}>
+        <span className="rt-messageLabel">{senderId}</span>
+        <div>
+          <span className="rt-message-text">{message}</span>{' '}
+          <span className="rt-messageLabel">
+            {sentAt ? formatTime(sentAt) : '???'}
+          </span>
+        </div>
+      </div>
+    </div>
+  )
+}