+++ website/src/components/demo/index.ts
@@ -1,21 +1,23 @@
+import { ChatDemo } from './ChatDemo'
 import { ConnectedUsersDemo } from './ConnectedUsersDemo'
 import CountButtonTogether from './CountButtonTogether'
 import HeroDemo from './HeroDemo'
 import { HoverHighlighterDemo } from './HoverHighlighterDemo'
 import { MeditationBell } from './MeditationBell'
 import { SessionManagerDemo } from './SessionManagerDemo'
+import { UseChatDemo } from './UseChatDemo'
 import { UseStateTogetherWPUVDemo } from './UseStateTogetherWPUVDemo'
-
 export * from './antdesign'
 export * from './primereact'
 
 export {
+  ChatDemo,
   ConnectedUsersDemo,
   CountButtonTogether,
   HeroDemo,
   HoverHighlighterDemo,
   MeditationBell,
   SessionManagerDemo,
-  UseStateTogetherWPUVDemo
+  UseChatDemo,
+  UseStateTogetherWPUVDemo,
 }
-