+++ website/src/pages/Documentation/DocumentationDemo.tsx
@@ -26,8 +26,9 @@ interface DocumentationDemoProps {
   url: string
   session1?: SessionParams | null
   session2?: SessionParams | null
+  aspectRatio?: string
 }
-export default function DocumentationDemo({ url, session1, session2 }: DocumentationDemoProps) {
+export default function DocumentationDemo({ url, session1, session2, aspectRatio }: DocumentationDemoProps) {
   const demoPath = `/demos/${url}`
   if (session1 === undefined) session1 = defaultSession
   if (session2 === undefined) session2 = session1
@@ -36,10 +37,10 @@ export default function DocumentationDemo({ url, session1, session2 }: Documenta
 
   return (
     <div className='w-full flex items-center flex-wrap gap-3 justify-center bg-white-100'>
-      <DocumentDemoContainer labelText='User 1'>
+      <DocumentDemoContainer labelText='User 1' aspectRatio={aspectRatio}>
         <Iframe url={url1} height='100%' width='100%' />
       </DocumentDemoContainer>
-      <DocumentDemoContainer labelText='User 2'>
+      <DocumentDemoContainer labelText='User 2' aspectRatio={aspectRatio}>
         <Iframe url={url2} height='100%' width='100%' />
       </DocumentDemoContainer>
     </div>