+++ website/src/utils/codeeditor/typescript/template/index.tsx
@@ -8,9 +8,8 @@ ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
     <ReactTogether
       sessionParams={{
-        {
-          // appId: import.meta.env.VITE_APP_ID,
-          // apiKey: import.meta.env.VITE_API_KEY,
+        appId: import.meta.env.VITE_APP_ID,
+        apiKey: import.meta.env.VITE_API_KEY,
         name: 'test',
         password: 'test',
       }}