+++ packages/react-together/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.3.1
+`2025-01-09`
+
+ - Added `omitMyValue` option flag in `useStateTogetherWithPerUserValues`
+ - Added `useChat` hook
+ - Added `<Chat/>` component
+
 ## 0.3.0
 `2024-12-13`
 
@@ -68,4 +75,4 @@
  - Added `useReactTogetherContext` hook
  - Added `<ConnectedViews/>` component
  - Added `<PresenceDiv/>` component
- - Added `<ReactTogether/>` context provider`
+ - Added `<ReactTogether/>` context provider